feat: implement market-data service with IBKR client, REST APIs, and caching
Phase 2 of options-trading merger:
- IBKR client layer: IbkrClient interface, FakeIbkrClient (dev), ConnectionManager, ContractResolver, SubscriptionManager
- Processing layer: QuoteNormalizer, OptionQuoteNormalizer, GreeksCalculator (Black-Scholes), OptionsChainBuilder
- Distribution: QuoteCacheService (Redis, 5s quotes / 30s chains)
- REST APIs: GET /api/v1/quotes/{symbol}, GET /api/v1/chains/{underlying}[/greeks], GET /api/v1/iv/{ticker}
- JPA entities and repositories for market_data schema
- Fixed V1 migration (con_id INTEGER, not BIGINT)

Verified: service boots, health UP, quote and chain endpoints return data with Greeks

Author: Saurabh Bilakhia Dawson <sbilakhia@dawsonpartners.com>

+++ backend/market-data/src/main/resources/db/migration/V1__market_data_schema.sql
@@ -52,7 +52,7 @@ CREATE INDEX idx_iv_observations_ticker_date
 CREATE TABLE market_data.contract_cache (
     id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     symbol VARCHAR(30) NOT NULL,
-    con_id BIGINT NOT NULL,
+    con_id INTEGER NOT NULL,
     sec_type VARCHAR(10) NOT NULL,
     expiry DATE,
     strike NUMERIC(12, 4),
